[IR] Treat all rescuable exception-raising instructions as hard barriers for local-var load/store placements.  Not possible to guarantee correctness otherwise.
Also identified possible "bug" in LVA for vars that cross eval boundaries
that affects the local-var load/store placement pass but which may not
affect Dead Code Elimination. To be investigated.

With a minor fix to TemporaryVariable.retrieve (which I am not
committing yet), Rubyspecs is green in 1.8 and 1.9 modes on this
pass now.  Also green with DeadCodeElimination is run before this pass.

The TemporaryVariable.retrieve fix has got to do with returning Nil if
a tmp var is undefined -- this is similar to DynamicScope.getValueOrNil
API call which becomes necessary for tmp vars as well when local vars
are treated like tmp vars with stores/loads to move them to/from
DynamicScope.  I want to look at ways to canonicalize the IR code to
initialize use-before-def locals to nil.