Fix recursiveKey usage across multiple threads (JRUBY-6621)
Instead of calling recursiveKey.set when the runtime is initialized,
we implement getInitialValue of ThreadLocal to supply the default
value when recursiveKey.get() is called for the first time on a given
thread.

This avoids an NPE when a runtime is initialized on one thread and
subsequently used on another thread under 1.9.