Remove all places where we unconditionally log exceptions.
Several locations in the embedding APIs unconditionally log
exceptions that are raised to the error stream configured in the
container. They then proceed to throw either the actual exception
or a wrapper, so the user ends up handling it anyway. This causes
expected exceptions to create unnecessary log noise and causes
unexpected exceptions to log twice as much information if they
are not handled by the user.

I have removed all places where exceptions are unconditionally
logged and then subsequently thrown in some form anyway.

Fixes #145.