Partial solution for #162.
This clears the ThreadLocal objects out, so they can GC, but it
causes the referenced objects to linger because ThreadLocal's weak
referencing still requires a few GC cycles. The better fix will be
to not use Ruby objects in the thread-locals, but that requires a
reimplementation of the recursion logic.