Stop the JITCompiler's executor during Ruby tearDown
This stops the JITCompiler's executor when tearing down a Ruby runtime
so that JITCompiler threads don't continue running
indefinitely. Before this change JITCompiler was leaking two threads
for every Ruby that gets created and subsequently torn down,
eventually causing problems in applications that have the potential to
create and destroy a lot of Ruby runtimes - like when running inside
TorqueBox or other web servers.