Fix compilation failures in spec runs.
The logic to get the class or module against which to run the body
was being emitted into the same Java method as the body. If it
accessed local variables or other context from the surrounding
body, it would break. This caused compilation failures on some
newer RubySpecs.

I moved the preparation of the class/module object outside the
compiled body, so it did not error when trying to access state
from the surrounding body.