Fix JRUBY-6619: NoMethodError 'upcase' on Java string
Well, this was a tricky one to find.

The problem stems from the actual Enumeration being walked. In
this case, when running under a Tomcat-based server (or perhaps
when running under a server that just uses some Apache libraries)
it is the NamesEnumerator class.

NamesEnumerator implements Enumeration<String>, which causes javac
to generate two nextElement methods: one with a String return and
one with an Object return. Issue number 1 is that JRuby does not
consider these to be overloads, and so it would randomly pick one
or the other depending on what order they were retrieved via Java
reflection. That was the cause of the variability.

JRuby converts return arguments using a cached "converter" that
it validates is still correct by confirming the object type being
returned matches the expected object type. If a method returns
String, we will use the cached String converter if the returned
object is also a String.

We also had code in place that attempted to find a "more public"
version of methods defined on non-public classes by searching the
public classes and interfaces above it.

In the case of NamesEnumerator, which is package-private, whenever
the String-returning method was encountered first we would proceed
to search for the Object version even though both were public
methods. But upon finding the Object version, we did not save it.
Instead, we only captured the converter based on an Object return
type and still dispatched using the String version.

So basically, we were calling the String version of the method but
using the Object converter to convert the result, which is why it
wrapped rather than converted.

I removed the public method search, since it wasn't actually
accomplishing anything. Both of the public methods in question
should dispatch ok without it, and this should avoid the broken
hybrid of String method with Object converter.