Fix structural problem with CSR request extension decoding.
The BounceCastle documented structure for PKCS10 request extensions is:

    Attributes ::= SET OF Attribute

    Attribute ::= SEQUENCE {
      type    ATTRIBUTE,
      values  SET SIZE(1..MAX) OF ATTRIBUTE.Type
    }

The code for extracting the ASN1 structures assumed Set, Set, Sequence - an
additional ASN1 Set implementation between the Attributes and the
Attribute itself.

This corrects that, by extracting only one level deep, resulting in the
correct behaviour.  It includes a test to avoid regressions.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>