Fix JRUBY-6717
jruby1.7.0preview1 doesn't work on Windows with jdk7

On some Windows systems (such as the reporter's) it appears that
the default Windows-1252 jcodings encoding does returns null for
getCharset(). I'm not sure why, but in this particular path I
believe it is safe to simply fall back on the Java default Charset
for transcoding.