Perf improvement for object duping.
When an object is dup'ed or cloned, all its instance variables
must be copied as well. The original logic did this by walking all
VariableAccessors stored in the object's real class, getting and
setting each value in turn. The new logic shortcuts this process
if the new realClass is the same as the old realClass, doing a
direct copy of the varTable.