Fix JRUBY-6788: Missing prompt in rails console
The issue here was that because we no longer launch child JRuby
processes within the same JVM, the console process that Rails
execs no longer had access to the TTY. This in turn caused it to
switch to "inspect mode" and to have no proper console prompts.

We had already started the move to a native exec by doing so on
Windows. My change expands true native exec to posix platforms and
adds a configuration option native.exec to disable the changed
behavior and keep the "subprocess" within the same JVM.

*Most* specs and tests pass, except for some new reopen failures
due to our reopen not actually reusing the same file descriptor
(it simulates reopening by swapping the NIO Channel being used by
the given ChannelDescriptor instance. I don't think this
regression is avoidable.