Fix JRUBY-6553
Application continues running after throwing out of memory exception in Thread.

I'm going with what some might see as a dubious fix: using
java.lang.Thread.stop(Throwable) to propagate the error to the
main thread. My justification is as follows:

* This only applies to Error cases like memory or stack, which are
usually fatal anyway (and even when not fatal, are probably a
sign of a major problem).
* It also only happens with abort_on_exception=true, which nobody
sets unless they *intend* for the process to exit on Thread-borne
errors.

The effect of this change is that when abort_on_exception=true is
set, Java Errors will also propagate to the main thread, likely
terminating the process.