Fix JRUBY-6766
--profile slows down Array#join , which distorts profiling results

This turned out to be a really good find.

In order to avoid the cost of dispatching to methods like
respond_to? and method_missing for value coercion, we pre-cache
those methods in org.jruby.Ruby and use identity checks to
shortcut faster, non-dynamic logic. This works well during normal
execution, but under profiling all method objects are wrapped
with ProfilingDynamicMethod objects that do the profile gathering.
Because the identity of the original method then becomes obscured
by the wrapper (which is created on a per-class basis) we always
followed the slow path. The slow path in the case of
method_missing raises an exception, and the exception backtraces
were the source of profiling slowdown.

I modified DelegatingDynamicMethod, the superclass of
ProfilingDynamicMethod, to override equals() to consider only the
wrapped method and fixed all related == comparisons to use
equals(), ensuring that regardless of which wrapper was
encountered the identity of the method would not be obscured. This
returned profiled performance for e.g. Array#join to normal
levels.

I should note that this fix improves profiling performance for all
coercion paths and methods that do coercion, so it has far-
reaching implications.

Thanks for reporting this :)