Fix (?) JRUBY-6746
OpenSSL 'connect_nonblock' missing/broken in 1.6.7 and 1.7.0.preview1

I implemented both connect_nonblock and accept_nonblock by passing
a "blocking" flag to all methods that might block. The blocking
methods usually did so by looping or waiting for selection events.

The script given in the bug does now run to completion, but it
does not re-attempt the connect. We still need to test that the
socket is not in a bad state after a failed nonblocking connect,
which I would guess means getting MRI 1.9 SSL tests running.