Fix JRUBY-6731
Can't load class files in WARs compiled by warbler using jruby 1.7.0.preview1

When requiring a file with an explicit .class extension, assume
the .class filename is the only filename to be searched.

Because a require of .class almost always will indicate that the
user wants to load a specific precompiled Ruby file, there's no
good reason (I know of) to search for anything else.

The original logic did not work at all; require 'foo.class' would
search for 'foo.class.rb' and 'foo.class.class' (in that order)
but never 'foo.class'.

The claim is that this worked in 1.6.7.2, but I could not
reproduce that with a simple case. The logic involved does not
appear to have changed in a substantial way since well before
1.6.7.2, and in any case I believe the fix here is the right way
for it to work.