Fix JRUBY-6802 (and probably JRUBY-6803)
JSON generation for largish data structures v. slow under 1.7 (JDK 1.6)
Reading thousands of files significantly slower with JRuby 1.7

This isn't a complete fix, but it solves the most egregious perf
issue.

The transcoding logic in JRuby currently does Charset.forName()
lookups repeatedly to get the proper Java Charset for a given
jcodings Encoding. Unfortunately, the "ASCII-8BIT" name used by
Ruby (and jcodings) does not map to a Charset, resulting in
every transcoding operation triggering an exception from forName.
Short-circuiting this logic explicitly for ASCII-8BIT avoids that
issue.

We're aware that there are other encodings that will also trigger
exceptions, and that we're repeatedly looking up the same Charsets
over and over. We will fix that in separate commits. Because the
bulk of the performance issues in these two bugs are fixed by this
patch, I'm going to call them resolved.