Temporarily set invokedynamic to be off unless turned on.
Invokedynamic gives Ruby code a tremendous performance boost on
JRuby, but unfortunately the current support for it in OpenJDK is
lacking. There are crashes and errors that prevent us from being
able to reliably recommend using it. Because of that, we are
temporarily setting the default behavior to not use invokedynamic,
until there are publicly available builds of OpenJDK with the new
and improved invokedynamic compiler support.

You can still turn on invokedynamic by passing
-Xcompile.invokedynamic=true to JRuby, or passing
-Djruby.compile.invokedynamic=true to the JVM, or by putting
these flags into JRUBY_OPTS or JAVA_OPTS, respectively.

We are working with the OpenJDK developers to ensure the new indy
logic gets into a public release as soon as possible.