Fix JRUBY-6849
pp Object.new shows 0x0000000 instead of a unique number for each object.

MRI logic appears to want to simply reproduce their "hashy"
inspect value, but this doesn't work in JRuby due to the lack of
the "p" format for Array#unpack. Modified to use a JRuby ext to
get the identityHashCode of the object. It is not unique, but
neither is our inspect output.