Fix JRUBY-6106
JRuby 1.9 coverage library reports different results than ruby 1.9

I have a basic understanding now of how MRI handles 'coverage'.

Coverage support is turned on by loading the 'coverage' extension.
This sets a global "rb_coverages" to non-nil, as a Hash.

At parse time, files parsed are added to the coverages hash as
having 'nil' for all lines. Files parsed before the 'coverage' ext
is loaded are not added to this hash. This is the first behavioral
difference in JRuby, since we're just adding lines to coverage for
any file+line encountered, regardless of whether the file was
loaded before or after 'coverage' ext loaded. See parse.y,
"coverage" function, called from yycompile0 to set up the array
for that file.

At compile time (after parse) code lines get reset to zeros
whenever that line would potentially trigger a "line" event in
trace hooks. This is in compile.c in the ADD_TRACE macro.
Essentially, compile.c adds the tracing logic as it goes only to
the lines that should trigger a line event, and at the same time
zeros out that line in the coverages array.

The determination of a "code line" happens in compile.c in the
iseq_compile_each function, right at the top, where it determines
the line is a newline.

    if (node->flags & NODE_FL_NEWLINE) {
	ADD_TRACE(ret, nd_line(node), RUBY_EVENT_LINE);
    }

So for every actual NEWLINE node in the AST, the according line in
the coverages array is reset to zero.

Assuming our set of newline nodes matches MRI's (questionable), it
should basically mean we just need to do the same zeroing of the
according line when we construct a NewlineNode in our parser.

The changes I made:

* In ParserSupport.newline_node, call
ParserConfiguration.coverLine with the appropriate line number

* In ParserConfiguration.coverLine, maintain a growing Integer[]
as lines are encountered. Zero the line being covered.

* In Parser, at the end of parse associate the filename with that
Integer[] only if coverage is enabled.

* In CoverageData's LINE event hook, only update lines for files
that already have an associated Integer[].

This should basically mimic MRI's behavior, modulo any newline
nodes we're missing (which is an open question).

There are a few places outside newline_node where we create
newlines. I did not add coverage calls there, because I'm unsure
if they're intended to be visitable lines or if they're outside of
normal executable code.