Fix JRUBY-6568
A Proc on a symbol like (&:first) works in 1.8, but not in 1.9 mode

In 1.9 mode, we do a better job of yielding the *actual* argument
given ([0] in this case) rather than always doing the extraneous
boxing that 1.8 mode does. However, we did not implement those
yieldSpecific paths in the Symbol#to_proc proc, so they fell back
on the old 1.8 logic that did arg list unwrapping. Implementing
the specific paths appears to make 1.9 mode yield to procified
Symbols properly.