Improve interpreted DStr by avoiding transient strings.
For an expression like "foo#{i}bar" we sometimes would create the
"foo" and "bar" strings only to concatenate them. For interpreter,
this only occurred if encodings were different, which was rare. I
fix that in this commit to always use the "foo" and "bar" ByteList
directly.

Compiler still creates "foo" and "bar" all the time. That must be
fixed.