Fix JRUBY-6773
This is actually a bug in stdlib. ERB#run and ERB#result both use
TOPLEVEL_BINDING to evaluate the template, and since by default
the template is appending to a String in _erbout, concurrent
template evaluations could step on each other.

My fix has been submitted to MRI in the following bug:

https://bugs.ruby-lang.org/issues/7046

The patch has not been accepted, but in the interest of fixing
this for JRuby users we are going ahead with it in JRuby 1.7RC1.