Fix edge case where rescued Exceptions may be nil
If an inner rescue block causes a Java proxy or package to be looked
up then an outer rescue block was receiving a nil Exception object,
only when running in interpretered, non-IR mode. Running the test with
"-X+C" or "-X-CIR" did not reproduce the behavior.

This behavior was introduced by the commits 20632af6 and 7c3f6426,
which were cleaning up a hack around the way $! got set. The mentioned
hack was covering up this bug.