Temporary fix for #312
I am creating a new exception as MRI does, but deferring that
until the timeout actually fires. This means for the majority of
cases that don't timeout, no anonymous exception class is ever
created.

I still don't like that a timeout-heavy system would spin through
classes like this, so I'm going to come up with a better
mechanism (and probably submit something to MRI as well).