Fix bug in indy attr_reader binding.
First time encounter with an attribute reader for a variable that
has never, ever been written is to produce a "dummy" accessor.
This accessor is not normally considered valid, but indy logic
invalidates based on actual class rather than real class, so the
validity of "dummy" accessors is never considered. As a result,
the dummy gets cached for as long as the target class is static,
which can mean even after the variable is assigned it still
returns a nil result.

Modified attribute reader logic in indy binding to always request
a real attribute reader, never a dummy. This allows indy's
invalidation logic to work at the cost of potentially eagerly
allocating variable accessors that won't eventually be assigned
(which ought to be pretty rare).