Fix JRUBY-6917
"already initialized constant ClassReader" when embedding JRuby

We were picking up the non-mangled ASM from Hadoop for
ClassReader, but because Hadoop does not ship the ASM "util"
classes, the subsequent access of TraceClassVisitor failed. We
then fell back on mangled names, but ClassReader got reassigned.

I swapped the order in which we check to ensure we always get our
mangled ASM if it's present, only falling back on non-mangled
when we can't find ours.