Fix JRUBY-6947
Call to "free" wrapped up via FFI causes JVM to exit with a malloc error.

Late in 1.7 dev, all of pointer.rb got moved into Java code in
JRuby. Along the way, write_string gained additional behavior:
it started to write trailing null character along with the String
bytes. This is incompatible with stock FFI. It may be a good
addition to the API, as write_cstring or similar, but I don't
think this change can be made compatibly.

Reverted that part of write_string's behavior to just write the
String's raw bytes. I'd like to see write_cstring added for the
\0 behavior.