Improved support for Server Name Indication (SNI)
Java 7 and greater supports SNI when creating SSL Sockets.
JRuby's origional support performed unnecessary reverse DNS
lookups if the SSL destination was an IP address instead
of a hostname.  This patch is more inline with how MRI
works, as it requires the developer to explicitly
set a hostname, thus avoiding any reverse lookups.