Move jruby-openssl gem bits into JRuby proper.
In order to keep the jruby-openssl gem always based on the most
current code, I'm moving the gem bits into JRuby proper. This
allows them to use the same source as JRuby itself; all we need
to do to spin a new gem is make code change in JRuby and build a
new gem from gems/jruby-openssl.

I also moved the 'gem' dir for the jruby-jars gem under
gems/jruby-jars.