Workaround for JRUBY-6967
Equals operator is not symmetrical and throws exceptions for Java classes

When we bind methods for a given class, we first search to see
whether the same method is defined in a superclass. If it is, we
consider that method to not have a "local" definition, and so we
don't define it.

In addition, the aliasing for == only happens at the highest
levels of the hierarchy, so overridden Java equals methods never
bind a new == because they're never local (except on j.l.Object)
and we don't assign aliases if there's no local impl of the method
in question.

This workaround fixes the issue by always forcing equals to be
considered "local" and forcibly adding a "==" alias to it. Any
class that defines its own equals method will now also have "=="
pointing at it, avoiding the Comparable#== that triggered this
issue.

There still needs to be a reexamination of how we bind methods.
The "local" search was intended to find a superclass method in
cases where the subclass may be private, breaking dispatch against
its direct reflected method. However, it causes the troubling
side effect from this bug, where mixed-in modules will override
the superclass methods (like "==") with no way for the subclass
to define their own versions if simply overriding.