Set up small literal hashes and hash args using a single bucket.
For many small literal hashes, especially those used for calling
methods with hash args, allocating the default layout for a Ruby
Hash is terribly wasteful. The default setup uses a bucket count
of 11, which for a hash smaller than 11 wastes that many entries.
Since many literal hashes and hash args are as small as one or two
or three elements, I have modified literal hash creation to force
a single bucket with no rehashing. This reduces memory allocation
of a simple single hash-arg call from 6.4MB to 2.4MB and improves
speed about 10-15%. There are additional improvements possible,
such as not rechecking the bucket count for hashing the incoming
keys.