Forced all sockets to use SO_REUSEADDR
If you have to have a default for ALL sockets, the safer assumption is that
there's going to be a lot more activity than less, so let's make ALL the
sockets set-up with the SO_REUSEADDR set to 'true'. Ideally, we will make
this a system define check, and maybe even an option, but those can come later
after we've proven this helps things.