Avoid perf-damaging extend of WaitReadable and WaitWritable.
In 1.9 mode, SSLErrors are created and then extend either
WaitReadable or WaitWritable if the error reason was that a
nonblocking read or write would block. Because extend changes
class hierarchies and creates new types, performance is impacted.
This patch makes a change similar to EAGAIN readable/writable in
normal IO by creating SSLErrorReadable and SSLErrorWritable that
simply extend SSLError and include the related module exactly
once.