Fix JRUBY-6999
Encoding from Windows-1252 to UTF-8 results in transcoding error

The jcodings Encoding for Windows-1252 is a replication of the
ISO-8859-1 Encoding, which caused it to use the wrong Java Charset
it inherited from ISO-8859-1. I modified the logic that negotiates
what Charset to use to make an extra effort to get the Charset
that actually corresponds with the encoding. It's not a great hack
but it will go away as jcodings fixes up things in the future.