Made changes as per Charles' request on exceptions
Charles asked me to change the code to NOT generate a second BindException, but
to change the newErrnoEADDRFromBindException() method, so I did that and just
make a second, more general form of that method that took a context message
string that would be put at the end of the excpetion's message. The old version
doesn't have to worry about this, but in these two cases, we can put a lot more
information at the user's fingertips.