Fix JRUBY-6951
Unable to encrypt data and then decrypt it in MRI Ruby

It appears that we were not setting up the initial IV properly. In
MRI, if you specify parameters to #encrypt, it will initialize IV
using a substring of "OpenSSL for Ruby rulez!". We had two issues:

1. In some places, we used "OpenSSL for JRuby rulez!"
2. If the initialization did not happen in #encrypt, we tried
again to do it on first update.

My fix was to use a blank IV (all zeros) if none is set up in
does not appear to break any other tests we run.