Move jruby-openssl out of jruby.jar.
For JRuby 1.7.0 and 1.7.1, we shipped all of jruby-openssl and the
Bouncy Castle dependency inside jruby.jar. This worked as an all-
in-one setup, but failed in other ways:

* With the jossl code in jruby.jar, it is impossible to update it
using a gem

* With BC code in jruby.jar, it is both impossible to update and
it fails to work for scenarios that require signed security
provider jars.

The new setup builds jopenssl.jar into lib/ruby/shared, copies the
bc*.jar files to lib/ruby/shared, and incorporates the
bouncy-castle-java.rb file from the bouncy-castle-java gem. This
allows jruby-openssl to be shipped with JRuby, outside jruby.jar
to avoid those issues, and upgradable by jruby-openssl gem.

We probably need to adjust dist logic for this. I'll coordinate
with enebo.