Properly set up the toplevel scope under new logic. Fixes #440.
The toplevel scope for a precompiled script was getting set up
with the toplevel scope for the runtime as parent, rather than
no parent. This caused constant lookup to search Object too soon,
at the top of the lexical search. This in turn led to constants
in the hierarchy with names the same as constants in Object to be
inaccessible.