Disable direct method binding for profiled methods.
When profiling is active, all methods get wrapped with a
ProfilingDynamicMethod wrapper. This wrapper handles timing the
call and recording aggregate and stack-oriented times. It also
obscures the original method as far as the indy binding logic is
concerned, since it wraps all types of methods but its logic needs
to be part of the dispatch.

This patch was initiated by a resulting bug from the wrapping
where our indy logic assumed anything that wasn't a Java dispatch,
attr access, or Ruby method was a native method that did not need
a scope. However, in the case of profiling, all methods fell into
that basket, resulting in a Ruby method that needed scope getting
directly bound without one, causing an NPE in ThreadContext scope-
pushing logic.

The fix here basically just disables direct binding for methods
wrapped in ProfilingDynamicMethod until we can add profiling
smarts to the indy binding logic.