Large rework of remaining test suites.
* Move all .rb-based suites out to rake
* Add 1.8 and 1.9 versions of relevant suites and make them pass
* Permute compiler flags over all .rb-based suites (excp: tracing)
* Remove Java-based test/unit runners and rename indices

Most developers should run `rake test19` as the shortest
day-to-day suite, `test18` for 1.8 features (fading away), or
`test` if changes might affect both modes. A more complete suite
is in the `:long` variants, and the full long suite with all
compiler permuations is in the `:all` targets. `ant test` and
related targets are still needed for JUnit, but they will get
moved to a rake target soon too.