[IR] Inline break handling into IR instruction stream.
Commit f9d0e85 introduced an inline break-handling pattern and
verified its feasibility.  This commit now transforms all calls
that might receive breaks into the following pattern in the
IR instruction stream

   try {
       call(... &blk)
   } catch (IRBreakJump e) {
       IRRuntimeHelper.handleBreak(e)
   }

Introduced a RuntimeHelperCall instruction for invoking a runtime
helper to handle the break in the rescue block.  In the future,
this instruction might expand to handle other runtime helpers in
the JIT.

The interpreter in turn no longers protects calls with a try-catch
but instead introduces an additional catch clause into the existing
RaiseException try-catch statement to handle IRBreakJumps.  In the
JVM bytecode (when JITted),  this effectively transforms into
additional exception table entries and is handled by the JVM
without any additional work to be do by the JIT.

This patch now eliminates all break-jump spec failures (previously
we had 4) since breaks from escaped blocks no longer receive
blanket-protection. They just bubble up to be trapped by a lambda
or get to the toplevel and halt execution.

TODO: Verify that if some piece of code has a LocalJumpError
rescue, that code runs properly (check if RubySpec has specs for
this already).

There is no performance penalty from this patch.