Remove assertions on instance var name.
I have run into two cases today where we are too strict about
instance variable names.

The first case I ran into was in Psych's accessing
internal instance variables with private_iv_get. We used
getInstanceVariable which attempts to enforce that all incoming
var names are valid instance variables. MRI does not enforce this
for C-level instance variable lookups, apparently.

The second is in krypt, accessing internal variables for similar
reasons.

Because of these two cases, and because we *know* there's
variables in the table with non-standard names, I'm removing the
assertions.

This allows JRuby + krypt to pass tests as before, even with
assertions turned on.