Switch JRuby's CPL license to EPL.
Governance of the CPL license was moved to the Eclipse Foundation
at some point in the past, and Eclipse blessed the EPL (Eclipse
Public License v1.0) as the successor to CPL. Because the CPL
contains a clause allowing transparent upgrades to future blessed
successor licenses, we can make the move to EPL without
re-contacting past contributors.

The change in going from CPL to EPL largely just removes some
punitive language that apparently made many companies nervous. All
other language in the license remains identical.

Several documents about Eclipse's move from CPL to EPL are here:

http://www.eclipse.org/legal/cpl2epl/

Note that Eclipse's move came before the blessing of EPL as the
upgrade path for CPL, so they had to do a bit more legal juggling.
With Eclipse now the administrator of CPL and with their blessing
of EPL as the upgrade path for CPL, all we need to do is make the
move.