Incorporate substantive code from 2.0 stdlib.
This commit only pulls in the parts of 2.0 stdlib that differ
functionality-wise from their 1.9.3 equivalents. When running in
2.0 mode, the 2.0 dir will be searched before the 1.9 dir,
picking up the differing files. We opted to do this to reduce the
amount of duplicate content in our repository.

JRuby 9000 will likely incorporate 2.0 stdlib en masse, and may
only be 2.0 compatible.

Note that 2.0 stdlib is *not* included in jruby-complete or
jruby-jars, since it is still an unsupported mode and we don't
want to bloat up those artifacts any more than necessary.