Implement Method#parameters for native methods.
In MRI, native methods will only present a list of :req for fixed-
arity methods or a single :rest for variable-arity methods. I have
duplicated this in JRuby, though I'd prefer to present the more
correct layout of required and optional args. I have filed a bug
with MRI to improve #parameters in this way:

https://bugs.ruby-lang.org/issues/8088

Fixes #583.